+++ back/config/routes.rb
@@ -127,9 +127,6 @@ Rails.application.routes.draw do
       resources :projects do
         resources :events, only: %i[new create]
         resources :projects_allowed_input_topics, only: [:index]
-        resources :topics, only: %i[index reorder] do
-          patch 'reorder', on: :member
-        end
         resources :phases, only: %i[index new create]
         resources :images, defaults: { container_type: 'Project' }
         resources :files, defaults: { container_type: 'Project' }
@@ -145,7 +142,7 @@ Rails.application.routes.draw do
       resources :projects_allowed_input_topics, only: [:index, :show, :reorder, :create, :destroy] do
         patch 'reorder', on: :member
       end
-      
+
       resources :admin_publications, only: %i[index show] do
         patch 'reorder', on: :member
         get 'status_counts', on: :collection