+++ back/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2022_01_26_110341) do
+ActiveRecord::Schema.define(version: 2022_02_07_103216) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -929,6 +929,15 @@ ActiveRecord::Schema.define(version: 2022_01_26_110341) do
     t.index ["topic_id"], name: "index_projects_allowed_input_topics_on_topic_id"
   end
 
+  create_table "projects_topics", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.uuid "topic_id", null: false
+    t.uuid "project_id", null: false
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.index ["project_id"], name: "index_projects_topics_on_project_id"
+    t.index ["topic_id"], name: "index_projects_topics_on_topic_id"
+  end
+
   create_table "public_api_api_clients", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "secret"
@@ -1207,6 +1216,8 @@ ActiveRecord::Schema.define(version: 2022_01_26_110341) do
   add_foreign_key "projects", "users", column: "default_assignee_id"
   add_foreign_key "projects_allowed_input_topics", "projects"
   add_foreign_key "projects_allowed_input_topics", "topics"
+  add_foreign_key "projects_topics", "projects"
+  add_foreign_key "projects_topics", "topics"
   add_foreign_key "public_api_api_clients", "tenants"
   add_foreign_key "spam_reports", "users"
   add_foreign_key "static_page_files", "static_pages"